Activate some clang-tidy checks (#1884)

+++ esphome/components/max7219digit/max7219digit.cpp
@@ -55,7 +55,7 @@ void MAX7219Component::dump_config() {
 }
 
 void MAX7219Component::loop() {
-  unsigned long now = millis();
+  uint32_t now = millis();
 
   // check if the buffer has shrunk past the current position since last update
   if ((this->max_displaybuffer_.size() >= this->old_buffer_size_ + 3) ||
