+++ esphome/components/slow_pwm/slow_pwm_output.cpp
@@ -12,7 +12,7 @@ void SlowPWMOutput::setup() {
 }
 
 void SlowPWMOutput::loop() {
-  unsigned long now = millis();
+  uint32_t now = millis();
   float scaled_state = this->state_ * this->period_;
 
   if (now - this->period_start_time_ > this->period_) {
